Running NTP checks against each noit is reasonable, but it would be good to be able to do all of this directly over port 43191.
Let's add a display of the time on the noit in the capabilities command.  This can be used by an API client as a poor-man's time check.  It will not be nearly as accurate as an NTP check, but we're looking for wild skew here anyway and this should do the trick.

closes #330


git-svn-id: https://labs.omniti.com/reconnoiter/trunk@1501 7e16921a-1b8a-68b2-a91d-dfb2e35db99f